[CONFIGURATION] * BUG: Do not compile database manager anymore, it has been deprecated. * BUG: Compile the gb.dbus component.
[INTERPRETER]
* NEW: Pointer is a real datatype now.
* NEW: Static arrays came back. They are now like real array classes, 
  except that they are not resizable.
* NEW: IsPointer() is a new function that tells if an expression is a 
  pointer.
* NEW: All XXXX?() function synonymous were removed. Use the IsXXXX() 
  version instead.
* NEW: Remove now useless source files.

[COMPILER]
* NEW: Static arrays came back, with actually the same syntax as before.

[EXAMPLES]
* BUG: Fix the examples according to the new array syntax.



git-svn-id: svn://localhost/gambas/trunk@2949 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_subr_extern.c
@@ -73,7 +73,7 @@ void SUBR_alloc(void)
   }
   
   RETURN->type = T_POINTER;
-  RETURN->_pointer.value = (intptr_t)ptr;
+  RETURN->_pointer.value = ptr;
   
   SUBR_LEAVE();
 }
@@ -116,7 +116,7 @@ void SUBR_realloc(void)
   REALLOC(&ptr, size * count, "SUBR_realloc");
   
   RETURN->type = T_POINTER;
-  RETURN->_pointer.value = (intptr_t)ptr;
+  RETURN->_pointer.value = ptr;
   
   SUBR_LEAVE();
 }
@@ -207,7 +207,7 @@ void SUBR_varptr(void)
 		THROW(E_ILLEGAL);
 
   RETURN->type = T_POINTER;
-  RETURN->_pointer.value = (intptr_t)ptr;
+  RETURN->_pointer.value = ptr;
 	
 	SUBR_LEAVE();
 }
